Don't assert on OUT being API handle in continuations
There's an assert to protect you from passing in API handles
as the `with` parameter to a continuation, because if you are
doing so you're likely losing access to that handle to free it.

However, the output cell of the frame is a legal `with`
parameter, and it may be an API value that is legitimate to
write to for the continuation (as legitimate as it would be
for a non-continuation, anyway.)

This relaxes the assert to allow an API handle if it's the
same as the frame output.