Add BLOCK! Isotope Behavior as Parameter Pack
The original prototype of multireturns was done as an enfix function
which could see the variables in the SET-BLOCK! on the left, and
as part of its execution would assign them.

That prototype was extended into being a feature of SET-BLOCK!
itself--though based on the evaluator doing more or less what the
enfix function had done.  This limited the ability to abstract (e.g.
to APPLY an multi-return function, or as much as put them in a GROUP!)

This is a reimagination of the multi-return mechanism to be based
on a more composable mechanism of returning an isotopic BLOCK!.  It
sacrifices the concept of "awareness" of what values are being
requested on the left, but provides enough benefits that it is
an unquestionable improvement.