Make DELIMIT react to BLANK! isotopes vs. literals
Because blanks would turn into nulls, DELIMIT had to hackily
look literally and see blanks in the feed before calling into
the evaluator to get them to act as spaces.

This did not give the code an opportunity to process the
blank, so any enfix operation that might have reacted to it
would not get the chance (and would be seen as missing its
left-hand argument)

Now BLANK! goes through an intermediate step on its way to
null decay, of being a BLANK! isotope.

But it's twisted so that the ~_~ quasiform does *not* produce
a blank, rather a pure null.  This frees up the`(@ ') operation
to give back a literal QUOTED! as it does for all types, while
only corrupting quasiforms...which helps the API, so it splices
~_~ in as a surrogate for nullptr vs. a hard-to-see single
apostrophe, that would call into question handling of all
QUOTED!s if any were corrupt.