Relax extant data stack assert during PROBE()
Because the stack can expand, it's dangerous to PUSH() while
a pointer to another cell that is PUSH()'d is being held onto
by code.  So there are systemic checks that keep a count of
how many such pointers are outstanding based on the StackValue()
wrapper class lifetime.

However it can be useful to be able to PROBE() during one of
these states, e.g.

    Derelativize(PUSH(), var, specifier);

If you're inside the implementation of Derelativize, you may
want to run the mold code...and this uses the data stack.

Most of the time the stack does not expand, so this reduces the
case to only asserting if the probe would actually cause a
crashing situation.