Separate "PATCH" into 3 distinct classes
When virtual binding experiments were first performed, there
was a single allocation type called a PATCH.  Sometimes this
would hold a variable declaration (e.g. a LET) and sometimes
it would hold a context that was inserted in lookup for
some (or all) of its variables.

When modules were shifted to "sea of words", the patch used
for LET variables seemed convenient to reuse, since words
would want to bind to these singular variables as they did
for LETs.  However, other differences made this confusing...
in particular confusing to debug.

Due to the difficulty of debugging a case involving virtual
binding patches, this goes ahead and separates them out into
three separate classes--which doesn't put much of an extra
burden on anything.