Workaround for console PRINT-RESULT bind bug
A strange method of printing values used by PRINT-RESULT
exposed a problem (one of many) with the current binding
model.  Binding needs a more complete review to see where it
will go next...so this just works around it, but keeps the
odd example as a memo of something to consider when rewriting.