Generalize QUASI! Scanning
The original ~xxx~ scanning was done via a quick hack and it
only supported words.  As the idea shifted from BAD-WORD! to
a generic QUASI! facility for all types, this needed to be
made more general.

So this hackily adds in that generalization--awaiting a better
implementation someday when the scanner gets rewritten.