Switch heavy null and void to ~[_]~ and ~[~]~
Before parameter packs, an empty splice as ~()~ was chosen as
being *sort of like* a void, in the sense that it had the
same behavior in the cases of things like APPEND.

When parameter packs arrived, the empty parameter pack ~[]~
was chosen as a "heavy null", while a parameter pack with nulls
in it (e.g. ~[_ ...]~ including ~[_]~) were decoded by the
THEN and ELSE as being synonymous with pure null intent.

This is changing, to where all parameter packs with nulls in
them are intepreted as "heavy" (compared to the pure NULL that
does not trigger a THEN).  With this change, it makes even
more sense to make the heavy version of void a parameter pack
as well.