ASSERT_SERIES() fix
Was not updated for the idea of REB_VOID and REB_ISOTOPE as
special VAL_TYPE() distinct from the heart byte.

While that idea may need review, and a return to stricter
handling in VAL_TYPE(), the assertions would still need to
allow the presence of voids and isotopes in frame arrays.