Update checks for Alloc_Tail_Array()
There was an assertion that when allocating from the tail of
the array that the cell was erased.  This is not generally
true--the space could be reclaimed and have old values in it.

This adds a new debug mode that will forcibly erase such
cells if desired, but not before checking that the cell is
either erased ("initable/fresh") or "writable".