Use PACK vs. multi-return FUNC in UPARSE
The multi-return functionality of FUNC is layered on top of
the isotopic block "pack" feature.

UPARSE has used that for combinators to return their REMAINDER
and their PENDING arrays.  But using FUNC's features directly
breaks some of the generality intended for passing on things
like the NONE result, which should be possible from a parser.

This uses PACK directly, allowing things like NONE (and other
packs) to be put normally into the output.