Add reb-c.h to %a-constants.c
The constants file is something of a leftover from R3-Alpha,
which tried to do minimal inclusions.  This meant it did not
include %sys-core.h.

A change to use nullptr instead of NULL discovered that it
thus didn't have the shim defined in C.  This adds reb-c.h, and
also goes ahead and takes out the rebol.h include file to
stay in the spirit of reduced dependencies for the file (if
that is important?)