Stop ANY and ALL from relying on STALE flag
This switches ANY and ALL from using the stale flag mechanic.
Only ALL really needed it (to preserve the old last result),
but ANY used it to avoid needing to clear the output cell.

In this reimplementation, ALL must copy a successful result
cell to the output on each cycle in order to save it from
being overwritten by potentially invisible subsequent
evaluations.