Macros for ME/MY, feed splicing for SHOVE
The SHOVE mechanic depended on NEXT_ARG_FROM_OUT, which is
being replaced by an enfix state byte in the action executor.

This changes to use machinery from MACRO, by actually rewriting
the feed.  This is more like what a generic mechanism in
usermode would look like for implementing the feature.

Because it splices material into the feed it runs afoul of a
binding issue that is causing other problems.  This adds another
hack to work around it until that can be redesigned--which stops
quoted arrays from getting virtually bound when seen inline
in a feed.

ME and MY are changed directly to use MACRO, in order to work
around binding complications with using SHOVE at this time.