Eliminate FEED_FLAG_NEXT_ARG_FROM_OUT flag
This gets rid of a problematic flag, by using a state byte in
the action dispatcher specifically for dealing with the mode
of acquiring arguments when the next one is supposed to come
from OUT.

The flag had problems being a feed flag, because the feed may
be shared among multiple frames and it was not clear which
one's "OUT" was being referred to.  Additionally, some cases
would check for it and others wouldn't.  Keeping the asserts
in line for it became fairly nonsensical over time--which
suggested it's a state that shouldn't exist.

Using a mix of techniques, the features depending on it were
split to using more narrow mechanisms.