Remove majority of stale bit functionality
Previously the evaluator would start off by marking all cells
with the stale bit (NODE_FLAG_FREE) in order to track that the output
had not been assigned.  This helped avoid spurious reads and
differentiated `(void) else [...]` from `(else [...])`.

But it required a bit being added and removed, so this replaces that
with the concept of "freshness" removing NODE_FLAG_NODE and
NODE_FLAG_CELL to get the effect with less overhead...preparing the
cell to be initialized and not requiring any masking out of bits.

So NODE_FLAG_FREE is now used in a more limited sense to get
unreadability for "poisoning"--a cell you can't read or write.