Require restart of Evaluator_Executor() state
When the Evaluator_Executor() was designed to do its own
stepping, it would reset its own state byte after every case
of it being set to something besides initial entry.

This was somewhat cluttering, and it also meant the byte was
being reset even when a frame was only going to be used once.
It also obscured what the state *had* been, and meant that a
frame might accidentally run twice.

There's also no intrinsic guarantee that a frame doesn't require
other setup besides a state byte to be reused.

So this changes to explicitly resetting the frame if it is
going to be used in a keepalive sense.