Use REB_XXX for evaluator executor state byte
It's multiply beneficial to have the state byte just convey the
actual type of the value being processed.  This is good for
debug information, it avoids having to come up with random
names for the continuation states, and it makes the value close
at hand for those handlers that want to use it without needing
to re-consult the value cell.