Skip over Just_Use_Out_Executor()
There was already a test to avoid clearing the state byte, so
if the test is paid for every loop it might as well just be
used to skip the executor.

This avoids having to make a state byte exception, and also
allows the Just_Use_Out_Executor() to have a stale cell if
it wants.