Eliminate entire MAYBE_STALE mechanic
It was important to distinguish when an output cell had been
written or not by the evaluator--for debugging purposes and
also for detecting an "end" to the left of something like
an enfix function, for instance (do [else [10]]).

A method was used that set a "stale bit" on the output cell.
This stale bit did not otherwise corrupt the contained value,
which permitted implementation of an invisibility feature by
recovering the stale value by unsetting the bit.

Some features could not work with this approach; e.g. an enfix
function could not consume its left argument and vanish, with
a reveal of whatever the value was before the left argument.
This is because the left argument had already overwritten the
output cell.  A different approach to invisibility which would
write to a temporary cell until an evaluation step was assured
to not be invisible wouldn't have the problem.

More features came about that were limited by the reliance of
invisibility on not writing the output cell--e.g. isotopic
value packs which contained invisibility, but the pack itself
would not be so it would overwrite the output.  Also, isotopic
objects which wanted to resolve themselves invisibly could not,
because their definition as an object wrote the output cell.

Ultimately this led to going ahead and using a temporary cell.
But also, the idea of using an additive state to indicate a
cell had not been written was replaced with a subtractive state;
the NODE_FLAG_NODE and NODE_FLAG_CELL were *removed*, which
meant that new cell headers could be OR'd on top without
needing to clear first.

This eliminates the remaining usages of the stale flag by
the trampoline and evaluator.