Synchronously raise errors in Array_Executor()
There was an idea that errors would be suppressed by the
array executor if void evaluations came along.  However, this
led to bad effects of running the next expression after an
error...it simply is too late to wait to see if something
became void.  Some other construct would be needed.