Add new handling for "heavy false"
In an upcoming change, ~false~ isotopes are just the normal
currency of falseness.  This means they can no longer serve
as an "ornery" surrogate for falseness.

This puts them in a pack.  It may not be necessary to do so,
but this allows trying out the idea that ELSE could be false
reactive as well.