Allow toplevel SET-TUPLE! in stripload trick
STRIPLOAD is a compatibility hack which lets code following
new rules be loaded in the mezzanine.

It was looking for patterns of top-level SET-WORD!s but did
not excempt patterns with dots in them, which would be TUPLE!.
There are now instances of SET-TUPLE! in the boot sequence
(due to SYS.UTIL.ENRESCUE:)

Filter dot-containing instances from the set of candidates
for top-level declarations.