Backwards-bootstrap-shim SPLIT-PATH, TRANSCODE
The removal of the ability to trigger multi-returns by
refinements mean that the bootstrap shim needs to manually
add those refinements so that code can be written in a style
that can be called be old and new.

This winds up being somewhat ugly and laborious, due to the
fact the bootstrap EXE cannot load SET-BLOCK!s.  So it has
to be ^META'd and unpacked by hand.