Separate out RESOLVE primitive from GET/SET
Because resolving the steps in a tuple is most of the cost
of pathing in a GET or a SET, they offered it as a multi-return.
But being multi-return affects the nature of GET and SET in
ways that are a poor fit for its applications, as it needs
fidelity in situations such as those where the thing being
gotten is actually a variable holding a parameter pack.

This creates a separate RESOLVE operation for the special cases
that need the steps.