PARSE- For Get Position of Incomplete PARSE
Originally, PARSE* did not require progress to the end of
the series.  It was a building block for PARSE that did have
that requirement.

However, the new defining difference of a lower level parse
from a higher level one is whether it returns a definitional
error in case of failure, or if it is wrapped at a higher level
that returns NULL or creates an object.

This brings back a PARSE- function that builds on top of
PARSE* with a composed rule using `<here>` that does not require
reaching the end and returns the position.