Workarounds for TCC's _Noreturn issues
TCC started warning about noreturn situations, and also added
the feature of _Noreturn from C11.  But there's no good way of
telling if TCC supports _Noreturn, outside of just detecting
C11.  Since the TCC boostrap attempts to work on older Linux
packaged distributions, this has workarounds for that...
and in the TCC extension if it sees a warning related to
returns it just ignores it.