First cut at ~true~ and ~false~ as isotopes, WIP
The long meditation over what "logic!" should be ultimately
settled that there should be no logic literal type, rather that
isotopic words of ~true~ and ~false~ would act truthy and falsey.

While this involves certain tradeoffs, it seems that these are in
alignment with the desire of users not to allow an odd logic
literal type to escape into representations.

(Significant reworking of the type system must be done in order
for this and other isotopic changes to cohere reasonably...so
many tests are failing, but the system can bootstrap and has
been hacked together to run until that can be done.)