Get rid of SCANT_EVALUATED_ISOTOPE flag
An ill-conceived idea made isotopic assignments like this from
a quasiform legal:

    >> x: ~(a b c)~
    == ~(a b c)~  ; isotope

While making indirect assignments error:

    >> spread [a b c]
    == ~(a b c)~  ; isotope

    >> x: spread [a b c]
    ** Error: can't assign isotope to X: without SET/ANY

This was hard to understand, and difficult to abstract.  A
function like EXPORT would be unable to easily emulate the
semantics.

A better strategy that gives full coverage for representation
in isotopic assignments is to use packs and quasiforms:

    >> [~var~]: ~[~(a b c)~]~
    == ~(a b c)~  ; isotope

The idea is further extended by using isotopic objects to wrap
up things that would be usually unassignable--like action
isotopes--and have those objects provide a methodized assignment
that does it quietly for these cases where generation of an
isotope is understood.