First boot of action isotopes (no UPARSE)
This is a first cut at trying to boot a system that follows
the rule that only isotopic actions in WORD! variables will
execute.

For the moment, it allows isotopic actions to be assigned via
SET-WORD! without complaint.  However, it's intended that this
would be illegal by default...and an action isotope would have
to be wrapped in an isotopic object that would coordinate to
do the assignment silently vs. error.  These isotopic objects
would be produced by RUNS.

It also makes any case of an isotopic action willing to decay
to a plain action.  This is not something that is actually
desirable...it's just a transitional state.

UPARSE is commented out for this boot, because it raises several
new questions.