Boot action isotopes including UPARSE
This takes the steps to finish up the booting of the system with
isotope actions running from WORD!, by including UPARSE.

The choice made is that the combinator MAP! stores non-isotopic
ACTION!, which then has to be RUN when fetched out of the map.
It has not yet been decided if maps will be more block like or
more object like, in terms of being allowed to store isotoppes.