No Automatic Parameter Decay of Action Isotopes
To get things booting, it was first tried that any parameter
which wasn't meta that received an action isotope (and didn't
have the ISOTOPES_OKAY flag) would get the reified version of
an action when passed the isotopic form.

If this behavior is a good idea at all, it should only be done
for functions that take *just* an ACTION!.  But even then it
may be presumptuous.

This shifts the responsibility to manually annotate when a
parameter can be an isotopic action.

It also takes away the automatic decay of isotope actions in
the COMPOSE function.