Move param_flags to Typeset's EXTRA field
At one point in time, a symbol had to be encoded in a typeset,
which was a non-bindable type.  So it had to use one of the
Any slots with a GC mark bit.

Things shuffled over time, and now the symbol is not stored in
the typeset at all--but rather in a separate keylist which is
an array of just symbol pointers.

This makes it more obvious that however typesets store their
actual data can be moved to the payload.