Change bootstrap detection of shimmed IMPORT
The previous technique used WORDS OF :IMPORT and FIND,
which is currently in flux in terms of whether WORDS OF
would allow you to ask that of an isotopic value.

This changes it to use TRAP of the act of specializing the
path :IMPORT/INTO, which will error if the /INTO refinement
is not available.  It works just as well and punts on the
question of what method WORDS OF would use.