Eliminate "spec" in DATATYPE! Cell
As preparation for eliminating the reified/lexical DATATYPE!,
this phases out the "spec" string in the cell for datatypes.

This feature is really a help feature, and the information from
the types.r file should probably be put somewhere in the
usermode help--optionally built in--vs. needing to get the
C code involved with it.