Eliminate "blank isotope" as prep for new null
The relationship between blank and null has been becoming more
hardened, where blanks evaluate to nulls.  But there was an
intermediate step where blanks turned into blank isotopes, which
would then turn into nulls in a later step.

But a new concept is to try saying that blank isotopes are
actually *what nulls are*, taking the mysterious non-visible
state and giving it entirely to void.

This takes one step toward that design point.