Transition NULL to be a blank isotope
At one point, it was deemed that being able to set variables
to null or to see null variables was best done with setting
them via an underscore.  So blanks would evaluate to null.

Due to the convergence of a few other ideas regarding how
nulls should be dealt with, the writing on the wall began to
be that null should actually be the isotopic state of blank.

So instead of NULL being some nebulous non-isotopic state
whose isotope form is VOID, this shuffled it around to where
VOID is the nebulous non-isotopic state, e.g. a single quote
mark is a quoted void...not a quoted null.

    >> '
    ; void

This meant that the isotopic state of void was available, and
could serve the role of the contents of an unset variable.
Here that is named NIHIL.