Fix memory corruption issue in sockets
Due to history, ports were using a BINARY! to store the state
information for a libuv socket.  However, if a libuv socket
is initialized and not shut down, then later polls for other
sockets can still trigger reading/writing the memory of the
idle socket--which may have had its BINARY! GC'd.  This uses
a HANDLE! so that if a port GCs with an outstanding libuv
TCP connection in it, the connection can be closed.