Make meta of null quasi blank, vs blank
When it was a somewhat tenuous link between NULL and BLANK,
the meta of null was set to be blank...not quasi-blank.

However, with null actually being a blank isotope, it makes
more sense and is more coherent for the meta to be a quasi
blank.  It is a little bit uglier, but the regularity is
worth it.