Introduce stable/unstable isotope distinction
This makes it impossible to put isotopic blocks, errors, or
objects into variables--they must be ^META'd.

A consequence of this change is that the previous trick for
writing NIHIL as a specialization of SET no longer works.
That was to use SET/ANY to assign an isotopic nihil, which
is taken by SET as a ^META parameter.