Encode null state as isotopic word ~null~
With the choice to use WORD! isotopes for special signaling
states of ~true~ and ~false~ vs. a dedicated logic type, the
idea of using the isotopic word ~null~ to represent null starts
to seem coherent.

This is also a natural outgrowth of using isotopic blank for
null.  But it's more legible...and it would be difficult to
train people to think that ~_~ is "technically a null".