Tweak to improve ^META parameter typecheck
Previously the system resorted to VAL_TYPE() of the heart of a
cell to check a meta parameter's type without changing it.

This switches to go ahead and unquote the value for the typecheck
and then put the quote level back on.

Also defers isotope type checking when calling continuations, so
that ~true~ / ~false~ / ~null~ isotopes can be honored.