Clearer workaround for extension map
BLANK! was being used in an odd way to work around a bug in the
null handling of a bootstrap executable on mac.  This workaround
wasn't as obvious as it could have been by using another type,
and also breaks if BLANK! is not considered falsey... and it is
not intended to be in an upcoming change.