Remove behavior that BLANK! evaluates to NULL
The experiment to evaluate blank to produce nulls was conceived
to reduce noise in the code.  But it was also predicated on this
removing object-level-display noise vs. just source noise, with
isotopic blank being thought of as the representation of null.

This was deemed less clear than the isotopic word of ~null~
being null, and likely to cause more confusion than it was
worth.  In light of this, blank was deemed that it would either
remain an inert type, or be returned to WORD!s where it could
be defined (as space, null, a module, etc).