Eliminate inner block handling in DELIMIT
There's a historically contentious behavior in "REJOIN"-type
functions which would automatically turn blocks into strings,
by taking the contents and smooshing them together non-reduced.

More often than not, this leads to gibberish.  Better strategies
now exist to make this dated concept unnecessary.

(Also, if there is block behavior there is a general rule that
BLANK! should act the same as an empty block.  So if block
behavior is removed, that makes it more sensible to allow the
blanks to have a dialected meaning, like intending a space.)

Fixes also some problems when rendering a null delimiter, since
nulls are now isotopes.