Blank handling in EMPTY?, FOR-EACH, MAP-EACH, SPREAD
With blank no longer falsey, its value as a placeholder type
comes not from its testing with IF...but from it not being
checked and yet acting with polymorphic grace with series
based values.

A period of time where BLANK! was acting as an "opt out" led
it to drop these polymorphisms.  That opting out was replaced
with the void intent, and so this imbues BLANK! again with the
purpose of not opting out but *opting in with nothing*...e.g.
acting more like an empty series than as a void.

An example to show the difference:

    >> for-each x [] [fail]
    ; void

    >> for each x _ [fail]
    ; void

    >> for-each x [a b c] [break]
    ; null

    >> for-each x void [fail]
    ; null