Unset map keys by setting to void, vs. null
Maps do not allow isotopes as either keys or values.

This avoids needing an exception to isotopic assignments to map
by making the way you unset keys in the map to void them.

It gives an extra level of assurance when assigning things in
a map that you are not accidentally assigning something you
thought was set but wasn't.