Tweak Just_Use_Out_Executor() optimization
Frames can be left on the stack--such as a frame for walking a
block doing a REDUCE--while calling another function (such as
a predicate).  These frames are given an executor which says
to just pass through the out value of their subframe while
being kept alive by the trampoline.

Rather than use `null` for the executor of these frames, a
no-op executor was used to be more clear of the intent.  But
there's no reason to actually call the function and go through
a trampoline iteration.  So there was an optimization for them.

That optimization had problems in the error case.  This updates
the optimization to force it through more code paths, but it
needs a more deep rethinking.