Tolerate REEVAL of isotopes
There was a problem with REEVAL taking isotopes, because the
evaluator does not expect to see isotopes.

It's not entirely clear what to do here, but the test that
was asserting was:

    true = reeval true

It doesn't seem like a bad thing to allow isotopes to reeval
to themselves.  But reeval needs to be rethought as part of a
general top-to-bottom thinking under a modern lens (like most
things in the system).