Tweaks so that SHOVE doesn't assert
At one time, an action wouldn't run enfixedly unless it was
bound to word.  Now plain ACTION! will execute enfixedly if it
is defined as enfix.

This hacks up a way to stop the SHOVE from crashing when it tried
to run enfixedly.  It's not fully fixed, but helps get through
the tests.