Add DEGRADE and CONCRETIZE, rethink DECAY
The term "decay" was used for turning quasiforms into isotopes,
and as an antonym of "reify".  But that is contentious with the
metaphor of isotopes: isotopic decay is about when an unstable
isotope (like a BLOCK! isotope) turns into its first element
when stored in variable.

It's unfortunate that REIFY doesn't have a good antonym besides
UNREIFY.  This tries out DEGRADE to see how it will work, and
makes DECAY serve the purpose of forcing things like isotopic
blocks to decay to the first value without a variable assignment.