Test fixes to reflect null as word isotope
For a little bit of time, null was the isotopic form of blank.
This meant "heavy null" was represented as isotopic ~[~null~]~

That was deemed too confusing, so it was changed to where it's
an isotope of the word null.  So the heavy form is ~[~null~]~

Other instances where blank isotopes were null ~_~ fixed too.