Rename REIFY method of isotopic objects to DECAY
The premise is that isotopic objects are unstable, and cannot
be stored in variables.  But they permit methodization, and
can contain handling for things like being used with THEN or
ELSE clauses (like "THEN-ables" from JavaScript).  However if
nothing expresses interest in its meta nature, it has to turn
into an ordinary value to be stored in a variable (or passed
as a normal argument to a function, which would mean it is
becoming a variable).

The name "reify" was given to this method.  However, it is
more accurate to call this behavior DECAY.  Whether it is
allowed to decay to another unstable isotope (such as another
isotopic object, pack, or error) is something that needs to
be explored...but it seems that at least packs should be
allowed, if not other things.