Non-variadic RETURN, QUIT, CONTINUE
It seemed like variadic behavior for RETURN and QUIT and
CONTINUE would make things cleaner.  But it introduced a dark
corner similar to JavaScript's problems with automatic semicolon
insertion.  People would become used to seeing RETURN or QUIT
or CONTINUE on their own line and assuming that meant the next
line wouldn't run at all, but it would be running and make
what was on the next line the argument.

Because of the diversity of intents for what RETURN with no
argument might mean, it's better to explicitly say RETURN NONE
or RETURN NULL or RETURN VOID, as each is a bit different.