Eliminate UNGET native
UNGET was introduced at a time when it was thought that GET-WORD!
would take on a different meaning, to actually both get and
UNMETA an item.  This was to work around issues where a meta
representation of an isotopic action! would be a regular action,
and you wanted to avoid running it.

Quasiforms solved this problem, to where all meta forms are
fetched through words in an "as-is" way.  So UNGET can just be
replaced with UNMETA.