Eliminate IMAGE! Datatype
Ren-C has not had a focus on GUI, and so the image datatype
was of limited use.

Additionally, serious semantic problems of thinking of images
as "series" were never answered by historical Rebol.

So in order to make room for more important datatypes, the
IMAGE! had been moved to act as an "extension type".  However,
as Ren-C faces some existential angst on whether the research
has an end point or not, it has become clear that grafting a
rich type system onto Rebol is a fool's errand.

In order to clear out flaky code and focus on a more narrow
and simple type system to make demo cases work (like UPARSE)
this eliminates the image!.  In browsers, images can be worked
with through the JavaScript interface and canvas.