Remove LIBRARY! Datatype
Library was based on dynamic loading of extensions.  This is
only a priority now in the context of webassembly, and would
require a significant investment.

The LIBRARY! datatype is the last of the extension types, and
moving it to an archive means the extension type machinery can
all be deleted.

  https://github.com/metaeducation/ren-c-library