Eliminate REB_CUSTOM type code
The concept of extensible types was ambitious but not really
articulated.  Rebol's historical fixation on 64 datatypes was
ingrained with no other real answer.

Ren-C added several basic structural types (like @[block] and
^(group), forcing things like image! and gob! into extensions.
These extensions cover cases that are not likely to be solved
in Ren-C's practical lifetime, so they have been removed.