Add TYPE-XXX! Lexical Types
This adds &[...], &(...), &... variants in order to attempt
to use them as a replacement for TYPESET!

The idea is that the block forms hold a list of alternates for
what would pass as a type, including functions.  Then the
group form would hold a list where all the conditions would
have to hold.