Typecheck errors give legal types, not illegal
Previously when an argument failed a typecheck it would not
print out the typeset of accepted types, because typesets
were very messy.  If the argument accepted ANY-VALUE! then
instead of being able to just print out [any-value!] as the
expected argument type, it would be a messy list of types.

Now that parameters encode an array of type constraints, it
is not only possible to give the constraints when a type
violation is encountered, it is preferable than implicating
the type of the actual argument passed.  That is because
some type constraints accept some values of a type and not
others, e.g. [even?] might reject an odd integer so saying
that the function does not accept integers is inaccurate.