Eliminate REB_NULL in favor of REB_ISOTOPE
This moves further toward legitimizing &isotope as one of
the answers to be given for the "kind of a value" as one
of the fundamental types, wrapping another type at quote
level -1.

Hence asking a null what it's "kind" is will now say that it
is an isotope, and further questions will need to be asked
to distinguish it more.