Add SWITCH/TYPE
With LOGIC! and NULL! not being datatypes but rather type
predicates, the pattern `switch type of x` cannot be used to
match such values.  This adds a /TYPE refinement to switch
so that instead you can write `switch/type x`.

(Some thought was given to if this should be a separate
construct like MATCH, but the name MATCH is taken for
a singular typecheck, e.g. `match integer! x`)