Some modernization of "BAD-WORD!" usages
Prior to the development of modern quasiforms and isotopes,
there were BAD-WORD!s that were kind of like "named UNSET!s"
from Rebol2 (e.g. legal in blocks but neither truthy nor
falsey, but could have any label).  This idea gave way to
the concept that everything you can put in a block is
"friendly" as well as truthy, and isotopic values and voids
are the only "ornery" values.

This tries to bring some reference points up to speed, so that
BAD-WORD! is chosen to specifically mean isotopic words (so
it would include ~null~ and ~true~ ~false~ isotopes, as well
as other word isotopes).  Maybe ISO-WORD! is a better term,
but there are few legitimate references anyway.