Speed up datatype checkers in TYPE_CHECK
This adds the same kind of acceleration used for typeset
checkers to the basic datatype checkers if they are used as
a type constraint.

Also removes the support for the quasi and quoted words in
type constraints, as the concept behind how that might work
has now changed.  It may be supported in the future.