Remove ! from types in %types.r
This is a first step toward getting the basic TYPE-WORD!s to
be like `&integer` instead of `&integer!`, by removing the
exclamation point in the %types.r.

The concept is that things like integer! will be constraints
based on (optimized) functions, and not equivalent to the
raw type like &integer.