Trust enum to initialize datatypes
In the original Rebol boot code, there was a list made of the
datatypes at the head of the boot block.  Ren-C forces the low
symbol IDs to line up with the datatypes, so there's no need
to pass in this list... it's simply understood that the REB_XXX
values up to REB_MAX correspond to symbol IDs with the same
value.  The list is redundant, so this drops it.