Separate builtin datatypes from Lib
This makes it so that the static list of built-in datatypes
is in a global table that is not tied to the definitions in
lib.  The goal is to make it so that INTEGER! and such can
be changed to type constraints like &(integer?), freeing up
the low ordered symbols for things like &integer.