Replace TO CHAR! and MAKE CHAR! with natives
With CHAR! becoming a type constraint, it's no longer legal to
use it with TO and MAKE.

This replaces the former functionality with natives that are
arguably clearer in their meaning.  e.g. previously it was not
clear that `TO CHAR! 1` should return either #1 or the control
code character for 1.  But CODEPOINT-TO-CHAR makes that more
obvious.