Fix set operation not running actions
The code for doing set operations on bitsets was built on top
of the code for set operations on BINARY!.  (This is a poor
implementation, which is better approached with something like
"Roaring Bitsets", but that work is incomplete.)

In delegating to the binary operation, it was making an API
call using an isotopic value.  The current behavior of using
isotopic values in the API is to degrade them into quasiforms,
which might not be the best idea to allow for ACTION!s.  But
because this was happening, the isotopic action was not
running in the delegation.