Fix QUIT and QUIT/WITH result meta
Right now, DO is written in terms of IMPORT...and IMPORT
gives back the result of evaluating a module's body as a
secondary parameter that has to be ^META in order to have
full coverage.  This is backwards: IMPORT should be written
in terms of DO and make the module a secondary output so
the main result can be an unstable isotope.

In the meantime, the handling of QUIT hadn't been set up to
return QUIT's argument in a meta form, so it was erroring.
This removes the UNMETA that was breaking this situation.