Sync EVALUATE/NEXT with passing a variable for pos
For a while, the absence or presence of a requested multi
return would trigger /NEXT vs. full evaluation behavior.
This was nixed when isotopic PACK! became the currency
of multi returns.

Because changing to use PACK! and other isotopic changes
were so disrupted, tests were not synchronized as the
experiment was pushed through.  This puts the tests for
EVALUATE back to pass a variable to hold the position, as
in traditional DO/NEXT.