Make (most) stable isotopes truthy
There was special handling to make isotopic groups and
actions truthy as far as ANY and ALL were concerned.  But
this didn't extend more generally to things like IF.

This stops testing of unstable isotopes for truthiness or
falseyness with an assert, and makes all the word isotopes
besides ~true~ ~false~ and ~null~ raise errors if tested.
But splices and activations are truthy.

Hence the special handling for ANY and ALL can be removed.