Don't use unstable isotopes in CONTINUE
Because it's possible to THROW unstable isotopes, there was
an idea that the unified function for trapping BREAK and
CONTINUE could distinguish a "breaking null" from a "pretend
the loop body finished with null" situation by making the
trap of CONTINUE represent nulls with heavy nulls.  This way
you wouldn't need a flag from the routine to say whether it
trapped a break vs. a continue--the contents of OUT would
not overlap.

This raised a problem with UNTIL where the CONTINUE result
was an unstable isotope that couldn't be tested for whether
it was truthy.  Having a boolean to track whether there is
a breaking condition turns out to be clearer.